refactor: Move files to components folder
Move components in src/app/records/** to components folder if they are used by a few pages
