+++ src/app/records/patient-consultation/page.tsx
@@ -1,13 +1,13 @@
-import { PatientInfoHeaderSection } from '@/components/records/PatientInfoHeaderSection';
+import { RecordConsultationTable } from '@/components/records/consultation/RecordConsultationTable';
+import { PatientInfoHeaderSection } from '@/components/records/patient/PatientInfoHeaderSection';
+import { PrescriptionTable } from '@/components/records/prescription/PrescriptionTable';
+import { PastVitalTable } from '@/components/records/vital/PastVitalTable';
 import { getConsultByVisitId } from '@/data/consult/getConsult';
 import { getPatientById } from '@/data/patient/getPatient';
 import { getVisitById } from '@/data/visit/getVisit';
 import { getVitalByVisit } from '@/data/vital/getVital';
 import { calculateDobDifference } from '@/types/Patient';
 import { EMPTY_VITAL } from '@/types/Vital';
-import { PrescriptionTable } from '../patient-record/PrescriptionTable';
-import { RecordConsultationTable } from '../patient-record/RecordConsultationTable';
-import { PastVitalTable } from '../patient-vitals/PastVitalTable';
 import { ConsultationForm } from './ConsultationForm';
 
 export default async function PatientConsultationPage({
