+++ src/app/records/patient-record/page.tsx
@@ -1,12 +1,12 @@
-import { PatientInfoDetailSection } from '@/components/records/PatientInfoDetailSection';
-import { PatientInfoHeaderSection } from '@/components/records/PatientInfoHeaderSection';
+import { RecordConsultationTable } from '@/components/records/consultation/RecordConsultationTable';
+import { PatientInfoDetailSection } from '@/components/records/patient/PatientInfoDetailSection';
+import { PatientInfoHeaderSection } from '@/components/records/patient/PatientInfoHeaderSection';
+import { PrescriptionTable } from '@/components/records/prescription/PrescriptionTable';
 import { getConsultByVisitId } from '@/data/consult/getConsult';
 import { getPatientById } from '@/data/patient/getPatient';
 import { getVitalByVisit } from '@/data/vital/getVital';
 import { EMPTY_VITAL } from '@/types/Vital';
 import { EditPatient } from './EditPatient';
-import { PrescriptionTable } from './PrescriptionTable';
-import { RecordConsultationTable } from './RecordConsultationTable';
 import { ViewVital } from './ViewVital';
 
 export default async function PatientRecordPage({
