+++ src/components/records/consultation/DiagnosesTable.tsx
