+++ src/components/records/patient/PatientInfo.tsx
